Fix issue with passing background data
Turns out the StorageService interface hadn't been loaded yet because
onBackgroundData happens before getInitialView... I would love to find a
better way to determine App context earlier in the life cycle... but for
now we will just load the StorageService as a Core service and depend on
the fact that none of the background objects have a dependency on it.

Next steps are to figure out how to push new views for the menu and
about pages.

Change-Id: Ic09d77ab7c30dc0120f2b37f5b0e6ba972dbc495