Significant Refactor of injection framework
Remove need to specify a separate Builder function for each binding.
Instead, bind the interface directly to a class def and unify the
constructor pattern for any class that has dependencies.

Remove the need to declare dependencies in the Module class. Instead
have classes implement a getDependencies function that can be called
without an object instance that can be queried by the binding to
determine what is actually required

Change-Id: I4fd01fa95d077314e38c932b35239b13d366b1c5