Add a Factory class to the Injection framework
Added a factory class that can be bound to an interface to provide
classes with a way to access the kernel without introducing a dependency
on the injection framework to the internals of the system.

Inject the factory into the remaining components that were still calling
new when pushing custom UI views and delegates.

This should be the last major step before calling the background update
changes and dependency injection good for testing.

Change-Id: Ie481533c29edb252e335dbdc3368be01b62532cd