Address Clipping issue on Fenix 6
Introduces new handling to dynamically determine optimal font size for
the device. Assumes "medium" unless that results in a timer offset
greater than 35% of the display size, then use "mild".

Dynamically determine the optimal location for the Timer Arc based on
the location of the Timer String using trig.

Fixes issue with determining the best location to write draw the timer
location, to use Graphics.getFontAscent() instead of the text height.

Issue: #20
Change-Id: I710174f6d603d97ac34dfc21c460a10060d453df